+++ golden-tests/GoldenTests.hs
@@ -1,25 +1,29 @@
+{-# LANGUAGE OverloadedStrings #-}
 module Main ( main ) where
 
 import Data.Algorithm.Diff
 import Data.Algorithm.DiffOutput
-import Data.Function ( on )
+import Data.Function ( on, (&) )
 import System.FilePath ( takeBaseName, replaceExtension )
 import Test.Tasty ( defaultMain, TestTree, testGroup )
-import Test.Tasty.Golden ( findByExtension, goldenVsString )
+import Test.Tasty.Golden ( findByExtension, goldenVsStringDiff )
 import Test.Tasty.Golden.Advanced ( goldenTest )
 
+import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as LBS
 import qualified Data.Text.Lazy as LazyText
 import qualified Data.Text.Lazy.Encoding as LazyText
 import qualified Data.Text.Lazy.IO as LazyText
+import qualified Data.Text.Prettyprint.Doc as Pretty
+import qualified Data.Text.Prettyprint.Doc.Render.Text as Pretty
 import qualified Dhall.Core
 import qualified Distribution.PackageDescription.Configuration as Cabal
-import qualified Distribution.PackageDescription.Parse as Cabal
+import qualified Distribution.PackageDescription.Parsec as Cabal
 import qualified Distribution.PackageDescription.PrettyPrint as Cabal
 import qualified Distribution.PackageDescription as Cabal
 import qualified Distribution.Verbosity as Cabal
 
-import CabalToDhall ( cabalToDhall )
+import CabalToDhall ( cabalToDhall, DhallLocation ( DhallLocation ) )
 import DhallToCabal ( dhallToCabal )
 
 
@@ -29,9 +33,55 @@ main =
   defaultMain =<< goldenTests
 
 
+preludeLocation :: Dhall.Core.Import
+preludeLocation =
+  Dhall.Core.Import
+    { Dhall.Core.importHashed =
+        Dhall.Core.ImportHashed
+          { Dhall.Core.hash =
+              Nothing
+          , Dhall.Core.importType =
+              Dhall.Core.Local
+                Dhall.Core.Parent
+                ( Dhall.Core.File
+                   ( Dhall.Core.Directory [ "dhall", ".." ] )
+                   "prelude.dhall"
+                )
+          }
+    , Dhall.Core.importMode =
+        Dhall.Core.Code
+    }
+
+
+typesLocation :: Dhall.Core.Import
+typesLocation =
+  Dhall.Core.Import
+    { Dhall.Core.importHashed =
+        Dhall.Core.ImportHashed
+          { Dhall.Core.hash =
+              Nothing
+          , Dhall.Core.importType =
+              Dhall.Core.Local
+                Dhall.Core.Parent
+                ( Dhall.Core.File
+                   ( Dhall.Core.Directory [ "dhall", ".." ] )
+                   "types.dhall"
+                )
+          }
+    , Dhall.Core.importMode =
+        Dhall.Core.Code
+    }
+
 
 goldenTests :: IO TestTree
 goldenTests = do
+  -- Note: must remain in sync with the layout options in
+  -- cabal-to-dhall/Main.hs, so that test output is easy to generate
+  -- at the command line.
+  let layoutOpts = Pretty.defaultLayoutOptions
+        { Pretty.layoutPageWidth = Pretty.AvailablePerLine 80 1.0 }
+      dhallLocation = DhallLocation preludeLocation typesLocation
+
   dhallFiles <-
     findByExtension [ ".dhall" ] "golden-tests/dhall-to-cabal"
   cabalFiles <-
@@ -61,11 +111,14 @@ goldenTests = do
           , let cabalFile = replaceExtension dhallFile ".cabal"
           ]
      , testGroup "cabal-to-dhall"
-         [ goldenVsString
+         [ goldenVsStringDiff
              ( takeBaseName cabalFile )
+             ( \ ref new -> [ "diff", "-u", ref, new ] )
              dhallFile
-             ( LazyText.readFile cabalFile
-                 >>= fmap ( LazyText.encodeUtf8 . Dhall.Core.pretty ) . cabalToDhall
+             ( BS.readFile cabalFile >>= cabalToDhall dhallLocation
+                 & fmap ( LazyText.encodeUtf8 . Pretty.renderLazy
+                        . Pretty.layoutSmart layoutOpts . Pretty.pretty
+                        )
              )
          | cabalFile <- cabalFiles
          , let dhallFile = replaceExtension cabalFile ".dhall"