+++ Changelog.md
@@ -25,6 +25,16 @@
 
 * `cabal-to-dhall` now pretty prints the resulting Dhall.
 
+* The signature of `CabalToDhall.cabalToDhall` has changed: it now takes the location
+  of the `prelude.dhall` and `types.dhall` to import as a parameter.
+
+* Upgrade to Cabal 2.2. This introduces SPDX license identifiers and Dhall 
+  functionality to manipulate them; see <golden-tests/dhall-to-cabal/SPDX.dhall>
+  for a (convoluted) demonstration.
+
+* `prelude.defaults.Executable` has lost its `main-is` field, as it
+  makes little sense to have an executable without it.
+
 
 ## 1.0.0.1 -- 2018-03-25